final take on set to particular height

+++ desk/desk.go
@@ -85,7 +85,7 @@ func ChangeToHeight(height float32) {
 	lock()
 	defer unlock()
 	var acceptableRange float32 = 0.75
-	for math.Abs(height - currentHeight) < acceptableRange {
+        for math.Abs(float64(height - currentHeight)) < float64(acceptableRange) {
 		acceptableRange *= 0.75
 	}
 	destLow := height - acceptableRange
